requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.34
pymysql==1.1.1

pydantic==2.8.2
pydantic-settings==2.4.0

passlib[bcrypt]==1.7.4
PyJWT==2.9.0

scikit-learn==1.5.1
numpy==2.0.1
joblib==1.4.2